header and landing page optimmized for mobile devices
