+++ src/styles/Launch.module.css
@@ -13,7 +13,7 @@
   flex: 1;
   display: flex;
   flex-direction: column;
-  align-items: left;
+  align-items: center;
 }
 .innerbox{
     width: 100%;
@@ -55,30 +55,9 @@
     border-width: 10px;
    
 }
-
-/* animation class and keyframes */
-.overflow-hidden {
- overflow: hidden;
-}
-
-.drop-in {
- animation: drop-in 2s ease 400ms backwards;
-}
-
-.drop-in-2 {
- animation: drop-in 2400ms ease 100ms backwards;
- 
-
-}
-
-@keyframes drop-in {
- from {
-  opacity: 0;
-  transform: translateY(-100px);
- }
- to {
-  opacity: 1;
-  transform: translate(0px);
- }
+@media only screen and (max-width: 600px) {
+  .topmobilemargin {
+    margin-top: 1rem !important;
+  }
 }
